# OpenAPI/Econom_main-openapi.yaml
openapi: "3.0.3"
info:
  title: "Econom API"
  version: "1.0.0"
servers:
  - url: "http://127.0.0.1:8080"
paths:
  /api/categories:
    get:
      operationId: "getCategories"
      responses:
        "200":
          description: "OK"
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/CategoryListDto"
  /api/categories/{category_id}:
    get:
      operationId: "getCategoryProducts"
      parameters:
        - name: "category_id"
          in: "path"
          required: true
          schema:
            type: "integer"
            format: "int64"
        - name: "page"
          in: "query"
          required: false
          schema:
            type: "integer"
            format: "int32"
            default: "1"
      responses:
        "200":
          description: "OK"
          content:
            '*/*':
              schema:
                type: "array"
                items:
                  $ref: "#/components/schemas/ProductDto"
  /api/products:
    get:
      operationId: "getProducts"
      parameters:
        - name: "page"
          in: "query"
          required: false
          schema:
            type: "integer"
            format: "int32"
            default: "1"
      responses:
        "200":
          description: "OK"
          content:
            '*/*':
              schema:
                type: "array"
                items:
                  $ref: "#/components/schemas/ProductDto"
  /api/products/{product_id}:
    get:
      operationId: "getProduct"
      parameters:
        - name: "product_id"
          in: "path"
          required: true
          schema:
            type: "integer"
            format: "int64"
        - name: "lon"
          in: "query"
          required: true
          schema:
            type: "string"
        - name: "lat"
          in: "query"
          required: true
          schema:
            type: "string"
      responses:
        "200":
          description: "OK"
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/ProductCost"
  /api/create-cart:
    get:
      operationId: "createCart"
      responses:
        "200":
          description: "OK"
          content:
            '*/*':
              schema:
                type: "string"
  /api/cart:
    get:
      operationId: "getCart"
      parameters:
        - name: "Authorization"
          in: "header"
          required: true
          schema:
            type: "string"
      responses:
        "200":
          description: "OK"
          content:
            '*/*':
              schema:
                type: "array"
                items:
                  $ref: "#/components/schemas/ShopCart"
  /api/cart/add:
    post:
      operationId: "addToCart"
      parameters:
        - name: "Authorization"
          in: "header"
          required: true
          schema:
            type: "string"
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CartControllerDto"
        required: true
      responses:
        "200":
          description: "OK"
          content:
            '*/*':
              schema:
                type: "array"
                items:
                  $ref: "#/components/schemas/ShopCart"
  /api/cart/delete:
    post:
      operationId: "deleteFromCart"
      parameters:
        - name: "Authorization"
          in: "header"
          required: true
          schema:
            type: "string"
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CartControllerDto"
        required: true
      responses:
        "200":
          description: "OK"
          content:
            '*/*':
              schema:
                type: "array"
                items:
                  $ref: "#/components/schemas/ShopCart"
components:
  schemas:
    CategoryListDto:
      type: "object"
      properties:
        id:
          type: "integer"
          format: "int64"
        name:
          type: "string"
        image_url:
          type: "string"
        isFinal:
          type: "boolean"
        children:
          type: "array"
          items:
            $ref: "#/components/schemas/CategoryListDto"
    ProductDto:
      type: "object"
      properties:
        id:
          type: "integer"
          format: "int64"
        name:
          type: "string"
        image_url:
          type: "string"
        category_id:
          type: "integer"
          format: "int64"
        link:
          type: "string"
    ShopCost:
      type: "object"
      properties:
        rus_name:
          type: "string"
        en_name:
          type: "string"
        price:
          type: "number"
          format: "double"
    ProductCost:
      type: "object"
      properties:
        id:
          type: "integer"
          format: "int64"
        name:
          type: "string"
        image_url:
          type: "string"
        priceList:
          type: "array"
          items:
            $ref: "#/components/schemas/ShopCost"
        best_cost:
          $ref: "#/components/schemas/ShopCost"
    CartItem:
      type: "object"
      properties:
        product_id:
          type: "integer"
          format: "int64"
        product_name:
          type: "string"
        image_url:
          type: "string"
        cost:
          $ref: "#/components/schemas/ShopCost"
        count:
          type: "integer"
          format: "int64"
    ShopCart:
      type: "object"
      properties:
        name:
          type: "string"
        products:
          type: "array"
          items:
            $ref: "#/components/schemas/CartItem"
        total:
          type: "number"
          format: "double"
    CartControllerDto:
      type: "object"
      properties:
        product_id:
          type: "integer"
          format: "int64"
        shop_name:
          type: "string"
        lon:
          type: "string"
        lat:
          type: "string"